src/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiohttp-cors==0.7.0
aiosignal==1.3.1
altair==5.4.0
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorful==0.5.6
contourpy==1.2.1
cycler==0.12.1
diffusers==0.30.0
distlib==0.3.8
fastapi==0.112.0
ffmpy==0.4.0
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
google-api-core==2.19.1
google-auth==2.33.0
googleapis-common-protos==1.63.2
gradio==4.41.0
gradio_client==1.3.0
grpcio==1.65.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdit-py-plugins==0.4.1
mdurl==0.1.2
memray==1.13.4
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
narwhals==1.4.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
opencensus==0.11.4
opencensus-context==0.1.3
orjson==3.10.7
pandas==2.2.2
pillow==10.4.0
prometheus_client==0.20.0
proto-plus==1.24.0
protobuf==5.27.3
py-spy==0.3.14
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
pyparsing==3.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
ray==2.34.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
ruff==0.5.7
safetensors==0.4.4
semantic-version==2.10.0
setuptools==72.1.0
shellingham==1.5.4
smart-open==7.0.4
sniffio==1.3.1
starlette==0.37.2
sympy==1.13.2
textual==0.76.0
tokenizers==0.19.1
tomlkit==0.12.0
torch==2.4.0
tqdm==4.66.5
transformers==4.44.0
triton==3.0.0
typer==0.12.3
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.19.0
virtualenv==20.26.3
watchfiles==0.23.0
websockets==11.0.3
wheel==0.44.0
wrapt==1.16.0
yarl==1.9.4
